third_party/WebKit/LayoutTests/editing/execCommand/insert-paragraph-in-list-item-anchor-crash.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id='sample' contenteditable>
<dl><a href="#"><b id="target">foo</b></a></dl>
</div>
<div id="log"></div>
<script>
test(function() {
    var sample = document.getElementById('sample');
    getSelection().collapse(document.getElementById('target'), 1);
    document.execCommand('InsertParagraph');
    // TODO(yosin): "InsertParagraph" should not insert DL around BR.
    assert_equals(sample.innerHTML.replace(/\n/g, ''), '<dl><b id="target"><a href="#">foo</a><a href="#"><dl><br></dl></a></b></dl>');
});
</script>
